requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Compress==1.13
Flask-Limiter==3.8.0

Werkzeug==2.3.7
SQLAlchemy==2.0.21
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3

Authlib==1.2.1
requests-oauthlib==1.3.1
python-dotenv==1.0.0
Flask-Mail==0.9.1

# DB
psycopg2-binary==2.9.10

# Optional (rate limiting memory fallback)
redis>=4.5

# Semantic Search
sentence-transformers>=2.2.0
numpy>=1.24.0